% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvisualization.R
\name{plotvisualization}
\alias{plotvisualization}
\title{Data Visualization using a pie chart with the plotly function}
\usage{
plotvisualization(druguse)
}
\arguments{
\item{druguse}{druguse$crack.frequency}
}
\description{
plotvisualization
}
\details{
The function needs the argument druguse
}
