% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reganalysis.R
\name{reg_analysis}
\alias{reg_analysis}
\title{reg_analysis function}
\usage{
reg_analysis(druguse)
}
\arguments{
\item{druguse}{data.frame}
}
\description{
Data Analisys: linear regression model to evaluete the "tranquilizer use" variation due to the "alcohol use" variation.

It's necessary to import other packages if the are not installed.
The function needs the argument druguse.
The output show some significant estimates that allow us to understand how the model is significative.
}
\author{
Francesca Grazia Radatti \email{francescaradatti97@gmail.com}
}
