% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regplot.R
\name{regplot}
\alias{regplot}
\title{regplot function}
\usage{
regplot(druguse)
}
\arguments{
\item{druguse}{data.frame}
}
\description{
The "regplot" function is an easy way to visualize the dependence of the tranquilizer frequency on the tranquilizer use.

The function needs the argument "druguse".
The output shows a graph built using the "ggplo2 package"
}
\author{
Francesca Grazia Radatti \email{francescaradatti97@gmail.com}
}
