% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvisualization.R
\name{plotvisualization}
\alias{plotvisualization}
\title{plotvisualization function}
\usage{
plotvisualization(druguse)
}
\arguments{
\item{druguse}{data.frame}
}
\value{
pie chart
}
\description{
The "plotvisualization" function is a smart way to represent how the drug use is dilueted by crack frequency using "plotly" package.
It's necessary to import other packages(for istance "magrittr")

The function needs the argument "druguse".

The output is a piechart that show percentage of the crack frequency across the various groups.
}
\author{
Francesca Grazia Radatti \email{francescaradatti97@gmail.com}
}
