% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regplot.R
\name{regplot}
\alias{regplot}
\title{data visualization 2: do some plots with some different characteristics}
\usage{
regplot(druguse)
}
\arguments{
\item{druguse}{data.frame}
}
\description{
data visualization 2: do some plots with some different characteristics
}
